debug infinite loops and fix LinExpr
These are mostly quality-of-life changes that fix a variety of subtle
interactions in NonLinExpr's __getattr__ and compare functions.

__getattr__ is a magic function that is called *not* whenever a class
performs object.attribute lookup, but whenever such a lookup throws
AttributeError. We need this function so NonLinExpr can "pretend" to be
a LinExpr. When an attribute that is not present in NonLinExpr is
requested, NonLinExpr can automatically compute a LinExpr and return the
corresponding attribute. The downside is that if __getattr__ itself
encounters AttributeError, the function gets stuck in an infinite loop.

The solution is to keep a set of attributes that are valid LinExpr
attributes, and throw AttributeError if the query isn't in the set.

__eq__, __ge__, and __le__ are magic functions which implement the
operators ==, >=, and <= respectively. We overwrite these functions
so that a NonLinExpr can again automatically become a LinExpr when
these operators are used to add a constraint to the linear program.

Python searches both sides of a comparison to find the proper magic
function. For example, if we're computing x == y, x.__eq__(y) is
checked, and if it returns NotImplemented, then y.__eq__(x) is computed.

The problem is if NonLinExpr's linear_form is invalid (i.e. not a
LinExpr), then its magic comparison functions won't work when comparing
against Var and return NotImplemented. Var's __eq__ function will then
be automatically called by Python. When Var is compared against a
LinExpr, the Var sends the comparison to the LinExpr using the "=="
operator instead of the __eq__ function. NonLinExpr gets the comparison
and again returns NotImplemented, Python checks the Var's __eq__
function, and gets stuck in an infinite loop.

The relevant lines are here:
https://github.com/coin-or/python-mip/blob/6c41f0d2311fce73877e549e70c6698f5345c7fd/mip/entities.py#L631-L632

A simplified view of the situation is as follows:
```
class NonLinExpr():

    def __eq__(self, other):
        return NotImplemented

class Var():

    def __eq__(self, other):
        return other == self

a, b = NonLinExpr(), Var()
print(a == b)
```

We can prevent the infinite loop by explicitly raising a
NotImplementedError if a comparison returns NotImplemented,
because we don't need the automatic switching behavior as
it should work properly without it.

The solution to the underlying issue was to properly return
a LinExpr, as the problem was that 0 times a LinExpr was
returning 0, which was not converted into a LinExpr.